T3C-784: Extract Report components for better organization
Refactor Report.tsx by extracting components into separate files:
- ReportLayout.tsx: Three-column layout with tab-aware outline visibility
- ReportHeader.tsx: Title, stats, summary, and overview components
- ReportToolbar.tsx: Expand/collapse buttons with tab awareness

Also move SubtopicItem wrapper logic from Topic.tsx into Subtopic.tsx,
creating a cleaner separation of concerns.

This reduces Report.tsx from ~1063 lines to ~679 lines.

Based on organizational improvements proposed by Brandon Perry in PR #390.
